Created the beginning of a menu building DSL
Why?
- We are going to need lots of menus for this application so having a DSL for building them will be very useful.

How?
The MenuCommand class contains everything needed to create a single command (e.g. help).
It takes a list of options and then passes them to OptionParser to be turned into a commandline interface.

MenuCommand::Command is a small class designed to hold basic information about each command.

The design of these classes allows us to use the syntax seen in the specs, which is a very succinct syntax.